requirements.txt
anthropic
edge-tts
moviepy==2.1.1
plotly
kaleido
Pillow
python-dotenv
numpy
mutagen
# openai-whisper  # 未來加字幕動畫時啟用：pip install openai-whisper
